[packaging] Replace rpm post scripts. JB#48950
pgrep -u root made sure that only the nemo-password-manager process
started by the root user would be sent the dbus kill signal. However,
nemo-password-manager normally starts as the root user, and it crashes
if attempted to run as the 'nemo' user with a sigabrt and:

[F] PasswordManager::PasswordManager:59 - Cannot register D-Bus service
at org.nemo.passwordmanager

Since this is basically a normal systemd service, fired off on dbus, it
makes more sense to use standard systemd/rpm maintainer scripts.

Thus, the pgrep hack is replaced by normal rpm macros.

This uses "BuildRequires: systemd" since on SFOS, it's the package that
holds the rpm macros, but if we split the macros in the future, that
should BuildRequire systemd-rpm-macros.